refactor: 패키지 구조 변경
- 클래스 기능에 맞게 패키지 이동
